% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_contrast.R
\name{mod_contrast_ui}
\alias{mod_contrast_ui}
\alias{mod_contrast_server}
\title{mod_contrast_ui and mod_contrast_server}
\usage{
mod_contrast_ui(id)

mod_contrast_server(input, output, session, image)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
