% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_histogram.R
\name{mod_histogram_ui}
\alias{mod_histogram_ui}
\alias{mod_histogram_server}
\title{mod_histogram_ui and mod_histogram_server}
\usage{
mod_histogram_ui(id)

mod_histogram_server(input, output, session, image)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
