% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_luminance.R
\name{mod_luminance_ui}
\alias{mod_luminance_ui}
\alias{mod_luminance_server}
\title{mod_luminance_ui and mod_luminance_server}
\usage{
mod_luminance_ui(id)

mod_luminance_server(input, output, session, image)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
