% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_image.R
\name{mod_sample_image_ui}
\alias{mod_sample_image_ui}
\alias{mod_sample_image_server}
\title{mod_sample_image_ui and mod_sample_image_server}
\usage{
mod_sample_image_ui(id)

mod_sample_image_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
